requirements.txt
python-dotenv
openai
tenacity

# OpenTelemetry packages
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging